pyproject.toml
[tool.poetry]
name = "sport-service"
version = "0.1.0"
description = ""
authors = ["Alex Martseniuk"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
psycopg2 = "^2.9.6"
django-db-logger = "^0.1.12"
drf-yasg = "^1.21.5"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
flake8-commas = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
